Reorganize controllers.
